+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `mlir::writer` module
 
+### Changed
+- Return `Value` from `InstBuilder::call`
+
 ### Fixed
 - Remove `%` from basic block names in `Block` formatting